+++ src/db/models/userBonus.js
@@ -22,11 +22,11 @@ module.exports = (sequelize, DataTypes) => {
     //   allowNull: false // Amount awarded to the user
     // },
     gcAmount: {
-      type: DataTypes.DOUBLE(10, 2),
+      type: DataTypes.DOUBLE,
       defaultValue: 0.0
     },
     scAmount: {
-      type: DataTypes.DOUBLE(10, 2),
+      type: DataTypes.DOUBLE,
       defaultValue: 0.0
     },
     wageredAmount: {
